Tech Debt: consolidate the two `processors` packages (#245)
### What

Consolidate the two `processors` packages

### Why

@aditya1702 and I ended up created both packages in different paths while working in parallel PRs. They should be the same package.
