+++ internal/indexer/processors/contract_operations.go
