+++ internal/indexer/processors/helpers.go
