+++ internal/indexer/processors/helpers_test.go
